requirements.txt
speedtest-cli==1.0.6
wheel==0.24.0
